enhanced the DatabasePage logic and style
Added a nav bar that contain the name of the project and the name of the current user also attached a menu with option to log out or go to the history page.
Redesigned the style of the databases to be  shown as a cards instead of columns.
Added the possibility for the user to create a new databases instead of uploading a existing one